% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetseq.R
\name{HetseqTest}
\alias{HetseqTest}
\title{Heterogeneity-seq: Testing for differential gene expression}
\usage{
HetseqTest(mat, A, B)
}
\arguments{
\item{mat}{Gene expression matrix of control cells}

\item{A}{Vector of cells (columns) in positive class}

\item{B}{Vector of cells (columns) in negative class}
}
\value{
Table of log2FC, p-values and adjusted p-values of differentially expressed genes
}
\description{
Testing for differential gene expression in predecessors of treated cells
}
