% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.doubleml}
\alias{plot.doubleml}
\title{Plot Heterogeneity-seq DoubleML Results}
\usage{
\method{plot}{doubleml}(
  table,
  highlights,
  p.cutoff = 0.05,
  est.cutoff = NULL,
  highlights.color = Seurat::DiscretePalette(36),
  highlight.cutoff = NULL,
  label.cutoff = 1.1,
  density.n = 500,
  point.scale = 0.5,
  xlab = "Estimate",
  ylab = bquote("-" ~ log[10] ~ FDR),
  linetype = "dashed"
)
}
\arguments{
\item{table}{Table from the Hetseq using the doubleML method.}

\item{highlights}{A list of gene vectors to highlight in the plot.}

\item{highlights.color}{A vector of colors for gene highlights.}

\item{label.cutoff}{Set a cutoff at which gene highlights will be labeled. Values > 1 will lead to no genes being labeled.}

\item{density.n}{Set granularity of 2d density color.}

\item{point.scale}{Set point size.}

\item{xlab}{Set label of the x-axis.}

\item{ylab}{Set label of the y-axis.}

\item{linetype}{Set the linetype of the p-value and estimate cutoff line.}

\item{highlights.cutoff}{Set a cutoff at which gene highlights should be set.}
}
\value{
ggplot object
}
\description{
This plotting functions creates a Vulcano Plot to visualize DoubleML Results.
}
\examples{
hetseq(method="doubleML", data, trajectories, score.name = "score", quantiles = c(0.25,0.75), compareGroups = c("Low", "High"))
plot.doubleml(res, highlights=list(c("GAPDH", "MYC", "ISG15")))
}
