% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mincostflow}
\alias{mincostflow}
\title{Min-Cost-Max-Flow for cellular trajectories}
\usage{
mincostflow(D.list, verbose = TRUE)
}
\arguments{
\item{D.list}{List of (pruned) distance matrices. Can be generated by the distmat function, see distmat example.}

\item{verbose}{Show verbose output.}
}
\description{
Applies Min-Cost-Max-Flow to calculate optimal trajectories from distance matrices.
}
\examples{
trajectories = mincostflow(D.list)
For full trajectory generation, see example of distmat function.
}
