% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mincostflow}
\alias{mincostflow}
\title{Min-Cost-Max-Flow for cellular trajectories}
\usage{
mincostflow(D.list, verbose = TRUE)
}
\arguments{
\item{D.list}{List of (pruned) distance matrices. Can be generated by the distmat function.}

\item{verbose}{Show verbose output.}
}
\value{
Matrix of cell-cell trajectories spanning all given timepoints.
}
\description{
Applies Min-Cost-Max-Flow to calculate optimal trajectories from distance matrices.
}
\examples{
\donttest{

# Full vignette available on https://grandr.erhard-lab.de/articles/web/hetseq.html

  mincostflow(D.list)
}
}
