% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotClassify}
\alias{PlotClassify}
\title{Plot Heterogeneity-seq Classifier Results}
\usage{
PlotClassify(
  table,
  highlights = NULL,
  highlights.color = Seurat::DiscretePalette(36),
  highlights.cutoff = NULL,
  label.cutoff = 1.1,
  density.n = 500,
  point.scale = 0.5,
  xlab = "AUC",
  ylab = bquote(log[2] ~ FC ~ (`0h`)),
  linetype = "dashed"
)
}
\arguments{
\item{table}{Table from HetSeq using the Classifier method.}

\item{highlights}{A list of gene vectors to highlight in the plot.}

\item{highlights.color}{A vector of colors for gene highlights.}

\item{highlights.cutoff}{Set a cutoff at which gene highlights should be set.}

\item{label.cutoff}{Set a cutoff at which gene highlights will be labeled. Values > 1 will lead to no genes being labeled.}

\item{density.n}{Set granularity of 2d density color.}

\item{point.scale}{Set point size.}

\item{xlab}{Set label of the x-axis.}

\item{ylab}{Set label of the y-axis.}

\item{linetype}{Set the linetype of the baseline AUC line.}
}
\value{
ggplot object
}
\description{
This plotting functions creates AUC Scatter plots to visualize Classifier Results.
}
\examples{
res=hetseq(method="classify", data, trajectories, score.name = "score", quantiles = c(0.25,0.75), compareGroups = c("Low", "High"), posClass=posClass)
plot.classify(res, highlights=list(c("GAPDH", "MYC", "ISG15")))
}
