% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{prune}
\alias{prune}
\title{Prune trajectories}
\usage{
prune(D.list, top.n = 10)
}
\arguments{
\item{D.list}{List of distance matrices. Can be generated by the distmat function, see distmat example.}

\item{top.n}{Prune trajectories to only top.n possible connections to optimize subsequent application of the mincostflow function.}
}
\description{
Prune trajectories down to top.n candidates to reduce runtime of subsequent mincostflow function.
}
\examples{
# For full trajectory generation, see example of distmat function.
D.list = prune(D.list)
}
